refactor(padi): L7 motion — fold terminal-workspace's node internals into @kolu/padi
Pure motion, zero logic changes. After W2.3 buried pulam, the node-only half
of @kolu/terminal-workspace (sensors · fold · fs/git endpoint) had exactly one
consumer — padi — while the browser-safe half (TerminalSnapshot schema +
agentProjection) is the shared vocabulary the client reads via kolu-common.
The package boundary cut nothing.

- MOVE into packages/padi/src/terminalWorkspace/: sensors.ts, fold.ts,
  endpoint.ts (+ their tests + gitRepo.testlib.ts). The moved files import the
  leaf's ./schema — the arrow points out (padi → leaf), never back.
- DELETE kavalChannels.ts — dead since pulam's burial (zero consumers).
- Shrink the leaf's exports to ./schema + ./agentProjection (drop ./endpoint
  and the bare `.` index; nobody imports them post-move) and prune its
  now-unused deps (@kolu/surface, @orpc/server, kaval, pino).
- padi gains kolu-github (sensors.ts needs it).

The moved files stay inside padi's hashed daemon closure (buildId.closure test
+ default.nix padiSrc), so padi's staleKey hash bumps once — one legitimate
padi restart on first deploy (re-seed, PTYs untouched), the precedented L3/L6
class.

Ledger: docs/atlas padi-cleanup.mdx §L7.
