+++ packages/padi/src/terminalWorkspace/sensors.test.ts
